REFACTOR: Change logo and Componentize DefaultHeader and DefaultFooter #38
