--- src/assets/Sass/default/_defaultHeader.scss
@@ -1,47 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Poppins:wght@400;500;700&display=swap');
-
-
-header{
-    font-family: 'Poppins';
-    height: 100px;
-    display: flex;
-    align-items: center;
-    padding: 40px 80px;
-    justify-content: space-between;
-    box-shadow: 0px 5px 7px 0px rgba(0, 0, 0, 0.25);
-    .header-leftSide{
-        display: flex;
-        align-items: center;
-        color: gray;
-        gap: 30px;
-
-        .links{
-            display: flex;
-            gap: 16px;
-
-            *{
-                font-size: 14px;
-                font-weight: 100;
-                color: gray;
-                text-decoration: none;
-            }
-        }
-        #logo{
-            width: 200px;
-            height: 36px;
-    }
-    }
-    .header-rightSide{
-        .icons{
-            display: flex;
-            align-items: center;
-            gap: 16px;
-
-            img{
-                width: 30px;
-                height: 30px;
-            }
-        }
-    }
-
-}
