+++ src/assets/Sass/default/footer/_FooterContainer.scss
@@ -1,0 +1,10 @@
+*{
+    font-family: Poppins;
+}
+footer{
+    box-shadow: 0px -1px 4px 0px rgba(0, 0, 0, 0.25);
+    padding: 40px 80px;
+    display: flex;
+    justify-content: space-between;
+
+}
