+++ src/assets/Sass/default/footer/_defaultFooter.scss
@@ -27,8 +27,7 @@ footer{
         gap: 48px;
         align-items: center;
             #logo{
-                width: 200px;
-                height: 36px;
+                width: 140px;
                 }
 
             .socialmedia-icons{
