+++ src/assets/Sass/default/footer/_logoAndIcons.scss
@@ -1,0 +1,16 @@
+.logo-icon{
+    .top-div{
+    display: flex;
+    gap: 48px;
+    align-items: center;
+        #logo{
+            width: 140px;
+            }
+
+        .socialmedia-icons{
+            display: flex;
+            gap: 16px;
+        }
+    }
+
+}
