+++ src/assets/Sass/default/header/_defaultHeader.scss
@@ -1,0 +1,12 @@
+@import url('https://fonts.googleapis.com/css2?family=Poppins:wght@400;500;700&display=swap');
+
+
+header{
+    font-family: 'Poppins';
+    height: 100px;
+    display: flex;
+    align-items: center;
+    padding: 40px 80px;
+    justify-content: space-between;
+    box-shadow: 0px 5px 7px 0px rgba(0, 0, 0, 0.25);
+}
