+++ src/assets/Sass/default/header/_rightSide.scss
@@ -1,0 +1,13 @@
+
+.header-rightSide{
+    .icons{
+        display: flex;
+        align-items: center;
+        gap: 16px;
+
+        img{
+            width: 30px;
+            height: 30px;
+        }
+    }
+}
