+++ src/components/default/TaskerLogo.vue
@@ -1,0 +1,3 @@
+<template>
+    <img src="../../assets/images/logo/tasker-logo.png" alt="">
+</template>
