+++ src/components/default/footer/DefaultFooter.vue
@@ -1,0 +1,9 @@
+<script setup>
+import FooterContainer from './FooterContainer.vue';
+</script>
+<template>
+  <FooterContainer/>
+</template>
+<style scoped>
+@import '../../../assets/Sass/default/footer/_defaultFooter.scss';
+</style>
