+++ src/components/default/footer/FooterContainer.vue
@@ -1,0 +1,13 @@
+<script setup>
+import ulFooter from './ulFooter.vue';
+import LogoAndIcons from './LogoAndIcons.vue';
+</script>
+<template>
+  <footer>
+    <LogoAndIcons/>
+    <ulFooter/>
+  </footer>
+</template>
+<style scoped>
+@import '../../../assets/Sass/default/footer/_FooterContainer.scss';
+</style>
