+++ src/components/default/footer/LogoAndIcons.vue
@@ -1,0 +1,19 @@
+<script setup>
+</script>
+<template>
+    <div class="logo-icon">
+      <div class="top-div">
+        <RouterLink to=""><img id="logo" src="../../../assets/images/logo/tasker-logo.png" alt=""></RouterLink>
+        <div class="socialmedia-icons">
+            <img src="../../../assets/images/instagram.png" alt="" />
+            <img src="../../../assets/images/github.png" alt="" />
+        </div>
+      </div>
+      <div class="bottom-div">
+        <p>2024 - Todos os direitos reservados</p>
+      </div>
+    </div>
+</template>
+<style scoped>
+@import '../../../assets/Sass/default/footer/_logoAndIcons.scss';
+</style>
