+++ src/components/default/header/DefaultHeader.vue
@@ -1,0 +1,9 @@
+<script setup>
+import headerContainer from'../../../components/default/header/headerContainer.vue';
+</script>
+<template>
+    <header>
+        <headerContainer/>
+    </header>
+</template>
+<style scoped></style>
