+++ src/components/default/header/LeftSide.vue
@@ -1,0 +1,17 @@
+<script setup>
+</script>
+
+<template>
+    <div class="header-leftSide">
+        <RouterLink to=""><img id="logo" src="../../../assets/images/logo/tasker-logo.png" alt=""></RouterLink>
+        <ul class="links">
+            <li><RouterLink to="/">Encontre trabalho</RouterLink></li>
+            <li><RouterLink to="/">Meus projetos</RouterLink></li>
+            <li><RouterLink to="/">Minhas finanças</RouterLink></li>
+        </ul>
+    </div>
+</template>
+
+<style scoped>
+@import '../../../assets/Sass/default/header/_leftSide.scss';
+</style>
