+++ src/components/default/header/dropdowns/MyProjects.vue
@@ -1,0 +1,9 @@
+<script setup>
+</script>
+
+<template>
+    <div></div>
+</template>
+
+<style scoped>
+</style>
