+++ src/components/default/header/headerContainer.vue
@@ -1,0 +1,14 @@
+<script setup>
+import LeftSide from './LeftSide.vue';
+import RightSide from './RightSide.vue';
+</script>
+<template>
+    <header>
+        <LeftSide/>
+        <RightSide/>
+    </header>
+</template>
+<style scoped>
+@import '../../../assets/Sass/default/header/_headerContainer.scss';
+
+</style>
