+++ src/views/global/HomeView.vue
@@ -3,7 +3,7 @@ import HeaderHome from '@/components/Home/HeaderHome.vue';
 import MainHome from '@/components/Home/MainHome.vue';
 import FreelaClient from '@/components/Home/FreelaClient.vue';
 import TalentsSkills from '@/components/Home/TalentsSkills.vue';
-import DefaultFooter from '@/components/default/DefaultFooter.vue'
+import DefaultFooter from '@/components/default/footer/DefaultFooter.vue'
 </script>
 
 <template>
